# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Dinanwuye API
  version: 0.1.0
  description: |
    Dinanwuye — Mobile-First Matchmaking Web App API contract.
    "Dinanwuye" = Igbo for "Husband and Wife".
    Base URL: https://api.dinanwuye.com/api/v1
    This is the initial spec stub for Phase 0 — endpoints will be
    generated per-service (NestJS Swagger + FastAPI) and merged here.
  contact:
    name: Dinanwuye Team
    email: admin@dinanwuye.com
  license:
    name: Proprietary
servers:
  - url: https://api.dinanwuye.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /safety/blocks:
    post:
      tags: [trust-safety]
      summary: Block a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockRequest'
      responses:
        '201':
          description: User blocked
  /safety/reports:
    post:
      tags: [trust-safety]
      summary: Report a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '201':
          description: Report submitted

  # ---------------- NOTIFICATIONS ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BlockRequest:
      type: object
      required: [target_id]
      properties:
        target_id:
          type: string
          format: uuid
        reason:
          type: string
    ReportRequest:
      type: object
      required: [target_id, reason]
      properties:
        target_id:
          type: string
          format: uuid
        reason:
          type: string
          enum: [harassment, fake_profile, inappropriate_content, other]
        details:
          type: string
    # ---------- Notifications ----------
...